simplify configuration for embed, image, video, placeholder
