+++ src/site/styles/layout/scaffold.scss
@@ -5,4 +5,18 @@
   box-sizing: border-box;
 }
 
+ul{
+  margin: 20px;
+  list-style: upper-roman;
+}
+
+h1, h2, h3{
+  margin: 0.5em 0;
+}
+
+blockquote{
+  padding: 0 2em;
+  margin: 20px 0px 4px 0px;
+}
+
