+++ src/tooltips.mdx
@@ -53,56 +53,6 @@ The default options for dante tooltip are:
   <Dante content={null} tooltips={[DanteInlineTooltipConfig({widget_options: {aa: "oo"}  })]} />
 </Playground>
 
-
-## Inline tooltip
-
-```
-{
-    ref: 'insert_tooltip',
-    component: DanteTooltip,
-    displayOnSelection: true,
-    selectionElements: [
-      "unstyled",
-      "blockquote",
-      "ordered-list",
-      "unordered-list",
-      "unordered-list-item",
-      "ordered-list-item",
-      "code-block",
-      'header-one',
-      'header-two',
-      'header-three',
-      'header-four',
-      'footer', 
-      'column',
-      'jumbo'
-    ],
-    widget_options: {
-      placeholder: "type a url",
-      
-      block_types: [
-        { label: 'p', style: 'unstyled',  icon: Icons.bold },
-        { label: 'h2', style: 'header-one', type: "block" , icon: Icons.h1 },
-        { label: 'h3', style: 'header-two', type: "block",  icon: Icons.h2 },
-        { label: 'h4', style: 'header-three', type: "block",  icon: Icons.h3 },
-
-        { type: "separator" },
-        { label: 'color', type: "color" },
-        { type: "link" },
-      
-        { label: 'blockquote', style: 'blockquote', type: "block", icon: Icons.blockquote },
-        { type: "separator" },
-        { label: 'insertunorderedlist', style: 'unordered-list-item', type: "block", icon: Icons.insertunorderedlist },
-        { label: 'insertorderedlist', style: 'ordered-list-item', type: "block", icon: Icons.insertunorderedlist },
-        { type: "separator" },
-        { label: 'code', style: 'code-block', type: "block",  icon: Icons.code },
-        { label: 'bold', style: 'BOLD', type: "inline", icon: Icons.bold },
-        { label: 'italic', style: 'ITALIC', type: "inline", icon: Icons.italic }
-      ]
-    }
-}
-
-```
 ## Default options
 
 <Playground>
